+++ udata_gouvfr/settings.py
@@ -1,0 +1,11 @@
+"""
+Default settings for udata-gouvfr
+"""
+
+# Frontend banner parameters
+BANNER_ACTIVATED = False
+BANNER_HTML_FILENAME = ''
+
+# Static pages from github repo
+PAGES_GH_REPO_NAME = 'etalab/datagouvfr-pages'
+PAGES_REPO_BRANCH = 'master'