+++ udata_gouvfr/tests/tests.py
@@ -25,11 +25,11 @@ from udata.tests.features.territories.test_territories_process import (
 from udata.utils import faker
 from udata.tests.helpers import assert200, assert404, assert_redirects, assert_equal_dates
 
-from .models import (
+from udata_gouvfr.models import (
     DATACONNEXIONS_5_CANDIDATE, DATACONNEXIONS_6_CANDIDATE,
     TERRITORY_DATASETS, OPENFIELD16, SPD
 )
-from .views import DATACONNEXIONS_5_CATEGORIES, DATACONNEXIONS_6_CATEGORIES
+from udata_gouvfr.views import DATACONNEXIONS_5_CATEGORIES, DATACONNEXIONS_6_CATEGORIES
 
 
 class GouvFrSettings(Testing):
@@ -132,7 +132,7 @@ class GouvFrHomeBlogTest:
 
     @pytest.fixture
     def home(self, mocker, client):
-        from . import theme
+        from udata_gouvfr import theme
 
         def home_client(blogpost):
             mocker.patch.object(theme, 'get_blog_post', return_value=blogpost)
@@ -189,7 +189,7 @@ class GetBlogPostMixin:
 
     @pytest.fixture
     def blogpost(self, app, rmock):
-        from .theme import get_blog_post
+        from udata_gouvfr.theme import get_blog_post
 
         def fixture(feed):
             if isinstance(feed, Exception):